data and state moved redux folder, refactoring browserrouter concept))
