+++ src/components/Profile/MyPosts/MyPosts.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 import Post from "./Post/Post";
 import { I18N } from "../../core/constants";
-import {addPost} from "../../core/mocks";
+import {addPost} from "../../redux/State";
 
 import s from "./MyPosts.module.css";
